requirements.txt
# Core dependencies
httpx>=0.27.0
anthropic>=0.39.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0

# Optional dependencies
pandas>=2.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
